split the I2C-specific tests into a Linux-only file
